userのusecase作成（リファクタ必要）

+++ pkg/application/user/bot_usecase.go
@@ -9,10 +9,10 @@ import (
 )
 
 // --- AppErrorType ---
-
 const (
 	BotDuplicateError      common.AppErrorType = "ボット情報が重複しています"
 	BotWebhookInvalidError common.AppErrorType = "ボットURLが無効です"
+	BotNotMineError        common.AppErrorType = "ボット情報は対象Userのものではありません"
 )
 
 // --- BotSearchInput ---
@@ -41,8 +41,13 @@ func NewBotSearchUseCase(
 
 // Execute Bot一覧取得を実行
 func (uc BotSearchUseCase) Execute(
+	context domain.UserAuthContext,
 	input BotSearchInput,
 ) (paginator query.BotSearchPaginatorDTO, err common.AppError) {
+	auth, e := context.UserAuth()
+	if e != nil {
+		return paginator, common.NewByError(e)
+	}
 
 	parameter := query.BotSearchParameterDTO{
 		Page:        input.Page,
@@ -51,7 +56,7 @@ func (uc BotSearchUseCase) Execute(
 		OrderByType: common.OrderByTypeDesc,
 	}
 
-	paginator, e := uc.query.Search(parameter)
+	paginator, e = uc.query.SearchByUserID(parameter, auth.ID())
 	if e != nil {
 		return paginator, common.NewByError(e)
 	}
@@ -76,18 +81,129 @@ func NewBotDetailUseCase(
 }
 
 // Execute Bot詳細取得を実行
-func (uc BotDetailUseCase) Execute(id uint) (detail query.BotDetailDTO, err common.AppError) {
+func (uc BotDetailUseCase) Execute(
+	context domain.UserAuthContext,
+	id uint,
+) (detail query.BotDetailDTO, err common.AppError) {
+	auth, e := context.UserAuth()
+	if e != nil {
+		return detail, common.NewByError(e)
+	}
+
 	findID, e := domain.NewBotID(id)
 	if e != nil {
 		return detail, common.NewByError(e)
 	}
 
-	detail, e = uc.query.FindByID(findID)
+	detail, e = uc.query.FindByIDAndUserID(findID, auth.ID())
 	if e != nil {
 		return detail, common.NewError(common.NotFoundDataError)
 	}
 
 	return detail, err
+}
+
+// --- BotStoreInput ---
+
+// BotStoreInput Bot追加DTO
+type BotStoreInput struct {
+	Name            string
+	AtatarFile      *multipart.FileHeader
+	AtatarDirectory string
+	Webhook         string
+	Active          bool
+}
+
+// --- BotCreateUseCase ---
+
+// BotStoreUseCase Bot追加ユースケース
+type BotStoreUseCase struct {
+	repository     domain.BotRepository
+	adapter        domain.DiscordWebhookCheckAdapter
+	fileRepository domain.BotAtatarImageRepository
+	service        domain.BotService
+}
+
+// NewBotStoreUseCase コンストラクタ
+func NewBotStoreUseCase(
+	repository domain.BotRepository,
+	adapter domain.DiscordWebhookCheckAdapter,
+	fileRepository domain.BotAtatarImageRepository,
+	service domain.BotService,
+) BotStoreUseCase {
+	return BotStoreUseCase{
+		repository,
+		adapter,
+		fileRepository,
+		service,
+	}
+}
+
+// Execute Bot追加を実行
+func (uc BotStoreUseCase) Execute(
+	context domain.UserAuthContext,
+	input BotStoreInput,
+) (id uint, err common.AppError) {
+	auth, e := context.UserAuth()
+	if e != nil {
+		return id, common.NewByError(e)
+	}
+
+	var avatar string
+
+	// アバターがアップロードされている場合
+	if input.AtatarFile != nil {
+		avatarEntity, e := domain.UploadBotAtatarImage(
+			input.AtatarFile,
+			input.AtatarDirectory,
+		)
+		if e != nil {
+			return id, common.NewByError(e)
+		}
+
+		avatarVO, e := uc.fileRepository.Store(avatarEntity)
+		avatar = avatarVO.Value()
+	}
+
+	nextID, e := uc.repository.NextIdentity()
+	if e != nil {
+		return id, common.NewByError(e)
+	}
+
+	entity, e := domain.CreateBot(
+		nextID.Value(),
+		input.Name,
+		avatar,
+		input.Webhook,
+		input.Active,
+		auth.ID(),
+	)
+	if e != nil {
+		return id, common.NewByError(e)
+	}
+
+	// ウェブフックの重複チェック
+	duplicate, e := uc.service.IsDuplicate(entity)
+	if e != nil {
+		return id, common.NewByError(e)
+	}
+	if duplicate {
+		return id, common.NewError(BotDuplicateError)
+	}
+
+	// ウェブフックの有効性チェック
+	ok, _ := uc.adapter.Check(entity.Webhook())
+	if !ok {
+		return id, common.NewError(BotWebhookInvalidError)
+	}
+
+	storeID, e := uc.repository.Store(entity)
+
+	if e != nil {
+		return id, common.NewByError(e)
+	}
+
+	return storeID.Value(), err
 }
 
 // --- BotUpdateInput ---
@@ -128,9 +244,15 @@ func NewBotUpdateUseCase(
 
 // Execute Bot更新を実行
 func (uc BotUpdateUseCase) Execute(
+	context domain.UserAuthContext,
 	id uint,
 	input BotUpdateInput,
 ) (err common.AppError) {
+	auth, e := context.UserAuth()
+	if e != nil {
+		return common.NewByError(e)
+	}
+
 	findID, e := domain.NewBotID(id)
 	if e != nil {
 		return common.NewByError(e)
@@ -139,6 +261,11 @@ func (uc BotUpdateUseCase) Execute(
 	entity, e := uc.repository.FindByID(findID)
 	if e != nil {
 		return common.NewByError(e)
+	}
+
+	// 自身のボットかどうか
+	if !entity.IsMine(auth.ID()) {
+		return common.NewError(BotNotMineError)
 	}
 
 	var avatar string
@@ -215,7 +342,14 @@ func NewBotDeleteUseCase(
 }
 
 // Execute Bot削除を実行
-func (uc BotDeleteUseCase) Execute(id uint) (err common.AppError) {
+func (uc BotDeleteUseCase) Execute(
+	context domain.UserAuthContext,
+	id uint,
+) (err common.AppError) {
+	auth, e := context.UserAuth()
+	if e != nil {
+		return common.NewByError(e)
+	}
 
 	deleteID, e := domain.NewBotID(id)
 	if e != nil {
@@ -225,6 +359,10 @@ func (uc BotDeleteUseCase) Execute(id uint) (err common.AppError) {
 	entity, e := uc.repository.FindByID(deleteID)
 	if e != nil {
 		return common.NewByError(e)
+	}
+
+	if !entity.IsMine(auth.ID()) {
+		return common.NewError(BotNotMineError)
 	}
 
 	e = uc.repository.Delete(deleteID)
